+++ __tests__/lexer/getParenStack.js
@@ -1,4 +1,4 @@
-const { getParenStack } = require("../src/utils/getParenStack.js");
+const { getParenStack } = require("../../src/lexer/getParenStack.js");
 
 describe("getParenStack", () => {
     it("should return an array of captures", () => {
