+++ __tests__/lexer/handleKleenePlus.js
@@ -1,20 +1,20 @@
-const { handleKleeneStar } = require("../src/utils/handleKleeneStar.js");
+const { handleKleenePlus } = require("../../src/lexer/handleKleenePlus.js");
 
-describe("handleKleeneStar", () => {
+describe("handleKleenePlus", () => {
     test("should throw an error if lastElement has a quantity other than 'exactlyOne'", () => {
         const lastElement = {
-            quantifier: "oneOrMore",
-            regex: "i+",
+            quantifier: "zeroOrMore",
+            regex: "i*",
             type: "literal",
             value: "i"
         };
         const nextChar = ".";
         function testError () {
-            handleKleeneStar(lastElement, nextChar, 0);
+            handleKleenePlus(lastElement, nextChar, 0);
         }
         expect(testError).toThrowError(new Error("Quantifier must follow an unquantified element"));
-        expect(lastElement.quantifier).toEqual("oneOrMore");
-        expect(lastElement.regex).toEqual("i+");
+        expect(lastElement.quantifier).toEqual("zeroOrMore");
+        expect(lastElement.regex).toEqual("i*");
     });
     test("should return the new index if greedy", () => {
         const lastElement = {
@@ -24,9 +24,9 @@ describe("handleKleeneStar", () => {
             value: "i"
         };
         const nextChar = ".";
-        expect(handleKleeneStar(lastElement, nextChar, 0)).toEqual(1);
-        expect(lastElement.quantifier).toEqual("zeroOrMore");
-        expect(lastElement.regex).toEqual("i*");
+        expect(handleKleenePlus(lastElement, nextChar, 0)).toEqual(1);
+        expect(lastElement.quantifier).toEqual("oneOrMore");
+        expect(lastElement.regex).toEqual("i+");
     });
     test("should return the new index if lazy", () => {
         const lastElement = {
@@ -36,8 +36,8 @@ describe("handleKleeneStar", () => {
             value: "i"
         };
         const nextChar = "?";
-        expect(handleKleeneStar(lastElement, nextChar, 0)).toEqual(2);
-        expect(lastElement.quantifier).toEqual("zeroOrMore-lazy");
-        expect(lastElement.regex).toEqual("i*?");
+        expect(handleKleenePlus(lastElement, nextChar, 0)).toEqual(2);
+        expect(lastElement.quantifier).toEqual("oneOrMore-lazy");
+        expect(lastElement.regex).toEqual("i+?");
     });
 });
