+++ src/lexer/getControlChar.js
