+++ src/lexer/getHexadecimal.js
