+++ src/lexer/getOctal.js
