+++ src/lexer/getUnicodeCharacter.js
