+++ src/lexer/handleEscapes.js
