+++ src/lexer/handleKleenePlus.js
