+++ src/lexer/handleKleeneStar.js
