+++ src/lexer/handleOptional.js
