+++ src/lexer/handlePeriod.js
